# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: NSW Agency Portal API
  description: |
    Backend API for the NSW Agency Portal. Provides endpoints for the workflow
    engine to inject applications for review, for agency officers to manage and
    review them, and for the frontend to handle file uploads and downloads.
  version: 1.0.0
servers:
  - url: /
    description: Local development server
...
paths:
...
  /api/v1/consignments/{consignmentId}:
    get:
      operationId: getConsignment
      summary: Get consignment
      security:
        - BearerAuth: []
      description: |
        Returns a single consignment summary for the officer portal. Display
        names (`consignmentName`, `consigneeName`) are fetched from NSW Core
        when available; they are omitted if Core is unreachable.
      tags:
        - Consignments
      parameters:
        - name: consignmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Consignment summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsignmentSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: invalid authentication token
    Forbidden:
      description: Authenticated but not authorised for this agency
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: access denied
    BadRequest:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Invalid request body: unexpected end of JSON input"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Application not found
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Failed to get application
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
...
    ConsignmentSummary:
      type: object
      description: Aggregated view of a consignment and its applications
      properties:
        consignmentId:
          type: string
          example: cefda05e-3071-4e94-b001-328094e570a7
        consignmentName:
          type: string
          description: Exporter/trader display name from NSW Core (omitted if unavailable or a placeholder such as N/A)
          example: Stay Naturals Private Limited
        consigneeName:
          type: string
          description: Consignee name from NSW Core (omitted if unavailable, GI has not run, or a placeholder such as N/A)
          example: Acme Imports Ltd
        updatedAt:
          type: string
          format: date-time
          description: Timestamp of the most recent update to this consignment
        status:
          type: string
          enum: [PENDING, FEEDBACK_REQUESTED, APPROVED, REJECTED, DONE]
          description: Status of the consignment (reflects the most recent application status)
        taskCount:
          type: integer
          description: Total number of applications in this consignment
          example: 3
...